Implements the Common Domain project.
The most significant changes include the update of the `.editorconfig` file to modify C# style preferences and diagnostic severity for certain rules. This change will affect the way code is written and analyzed in the project. Another major change is the addition of a new package reference to `MediatR` in the `Domain.csproj` file, which will introduce new functionalities to the project. The addition of several new files to the `Domain.Primitives` namespace is also noteworthy as it introduces new classes and interfaces related to domain-driven design.

1. The `.editorconfig` file was updated to change the C# style preferences and diagnostic severity for certain rules. The `csharp_prefer_simple_default_expression` and `csharp_style_prefer_null_check_over_type_check` were changed from `true` to `true:suggestion`, and the severity of the `SA1010` and `CA1030` diagnostics was set to `none`.

2. The `Domain.csproj` file was updated to include a new package reference to `MediatR` version `12.2.0`.

3. A new `GlobalUsings.cs` file was added with a global using directive for `MediatR`.

4. Several new files were added to the `Domain.Primitives` namespace, including `Entity.cs`, `IAuditable.cs`, `IDomainEvent.cs`, `IEntity.cs`, `IEntityId.cs`, and `ValueObject.cs`.

5. A new `SystemTimeProvider.cs` file was added to the `Domain.Time` namespace. This file defines a static class that provides a function for getting the current date and time.